test(adapta): isolate non-stream boundary fixture
